backend/pyproject.toml
[tool.poetry]
name = "ecocash"
version = "0.1.0"
description = "EcoCash Assistant Backend"
authors = ["Tyler Slaton <tyler@copilotkit.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
langgraph = "*"
langchain-openai = "^0.3.27"
langchain-anthropic = "^0.2.1"
langchain-google-genai = "*"
langchain = "^0.3.1"
langchain-core = "^0.3.25"
openai = "^1.51.0"
langchain-community = "^0.3.1"
copilotkit = "0.1.54"
uvicorn = "^0.31.0"
python-dotenv = "^1.0.1"
tavily-python = "^0.5.0"
html2text = "^2024.2.26"
googlemaps = "^4.10.0"
langgraph-cli = {extras = ["inmem"], version = "^0.1.64"}
langgraph-checkpoint-mongodb = "*"
pymongo = "*"
mcp = "*"
langchain-mcp-adapters = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
demo = "travel.demo:main"
